+++ src/components/recipeCreate/RecipeRadio.jsx
@@ -1,13 +1,13 @@
 const RecipeRadio = ({ label, value, register, config={}, onChange=null }) => {
 	return (
     <>
-      {/* <label>{ value }</label>
+      <label>{ value }</label>
       <input 
         type="radio" 
         value={ value }
         {...register( label, config )}
-        onChange={onChange} 
-      /> */}
+        onChange={ onChange } 
+      />
     </>
   )
 }
